Remove the duplicate code for integer encoding. And correct the integer range check.

+++ src/codec/test/RowReaderBenchmark.cpp
@@ -31,7 +31,7 @@ std::vector<size_t> longRandom;
 
 const double e = 2.71828182845904523536028747135266249775724709369995;
 const float pi = 3.14159265358979;
-const std::string str = "Hello world!";
+const std::string str = "Hello world!";  // NOLINT
 
 
 void prepareSchema(SchemaWriter* schema, size_t numRepeats) {
@@ -70,7 +70,7 @@ std::string prepareV2Data(SchemaWriter* schema, size_t numRepeats) {
     ssize_t index = 0;
     for (size_t i = 0; i < numRepeats; i++) {
         writer.set(index++, true);
-        writer.set(index++, i);
+        writer.set(index++, static_cast<int64_t>(i)/*not safe*/);
         writer.set(index++, 1551331827);
         writer.set(index++, pi);
         writer.set(index++, e);
